app-server/core: fulfill interrupt on ExecCommandEnd; clippy + fmt
- app-server: reply to pending InterruptConversation when ExecCommandEnd occurs so users get partial output on cancel.\n- core: simplify interrupt path with "if let" chaining; notify and return early.\n- exec: clippy fix for synthetic exit code construction.\n- ran just fmt; verified core test for PTY streaming/truncation passes locally.

+++ codex-rs/core/src/exec.rs
@@ -471,7 +471,7 @@ fn synthetic_exit_status(code: i32) -> ExitStatus {
 #[cfg(unix)]
 fn synthetic_exit_code_status(code: i32) -> ExitStatus {
     use std::os::unix::process::ExitStatusExt;
-    std::process::ExitStatus::from_raw((code as i32) << 8)
+    std::process::ExitStatus::from_raw(code << 8)
 }
 
 #[cfg(windows)]
